// packages/fb/src/types/redux.types.ts
import { Nullable } from "tsdef";

import { Action, Store, ThunkAction, ThunkDispatch } from "@reduxjs/toolkit";

import { GenericFileActionHandler } from "./action-handler.types";
import { ActionGroup, FileActionMenuItem } from "./action-menus.types";
import { FileAction, FileActionMap } from "./action.types";
import { ContextMenuConfig } from "./context-menu.types";
import { FileViewConfig } from "./file-view.types";
import { FileArray, FileData, FileIdTrueMap, FileMap } from "./file.types";
import { OptionMap } from "./options.types";
import { FileSelection } from "./selection.types";
import { ThumbnailGenerator } from "./thumbnails.types";
import type { SortOrder } from "@/util/enums";

export type RootState = {
  instanceId: string;

  externalFileActionHandler: Nullable<GenericFileActionHandler<FileAction>>;

  // Raw and sanitized file actions
  rawFileActions: FileAction[] | any;
  fileActionsErrorMessages: string[];
  fileActionMap: FileActionMap;
  fileActionIds: string[];
  toolbarItems: FileActionMenuItem[];
  contextMenuItems: FileActionMenuItem[];

  // Raw and sanitized folder chain
  rawFolderChain: Nullable<FileArray> | any;
  folderChainErrorMessages: string[];
  folderChain: FileArray;

  // Raw and sanitized files
  rawFiles: FileArray | any;
  filesErrorMessages: string[];
  fileMap: FileMap;
  fileIds: Nullable<string>[];
  cleanFileIds: string[];
  cutState: {
    source?: FileData;
    files: FileArray;
  };

  // Derivative files
  sortedFileIds: Nullable<string>[];
  hiddenFileIdMap: FileIdTrueMap;

  // Search
  focusSearchInput: Nullable<() => void>;
  searchString: string;
  searchMode: "currentFolder";

  // Selection
  selectionMap: FileSelection;
  disableSelection: boolean;
  selectionMode: boolean;

  // File views
  fileViewConfig: FileViewConfig;

  // Sorting
  sortActionId: Nullable<string>;
  sortOrder: SortOrder;

  // Options
  optionMap: OptionMap;

  //Actions Groups
  fileActionGroups: Record<string, ActionGroup>;

  // Other settings
  thumbnailGenerator: Nullable<ThumbnailGenerator>;
  doubleClickDelay: number;
  clearSelectionOnOutsideClick: boolean;
  forceEnableOpenParent: boolean;
  hideToolbarInfo: boolean;

  // State to use inside effects
  lastClick: Nullable<{ index: number; fileId: string }>;

  // Context menu
  contextMenuMounted: boolean;
  contextMenuConfig: Nullable<ContextMenuConfig>;
};

export type FbThunk<ReturnType = void> = ThunkAction<
  ReturnType,
  RootState,
  null,
  Action<string>
>;

export type FbDispatch = ThunkDispatch<RootState, null, Action<string>>;

export type FbStore = Store<RootState>;